break down interfaces
