+++ pkg/services/dcs/types.go
@@ -1,0 +1,13 @@
+package dcs
+
+import (
+	"code.cloudfoundry.org/lager"
+	"github.com/huaweicloud/huaweicloud-service-broker/pkg/config"
+)
+
+// DCSBroker define
+type DCSBroker struct {
+	CloudCredentials config.CloudCredentials
+	Catalog          config.Catalog
+	Logger           lager.Logger
+}
