+++ pkg/services/dms/provision.go
@@ -1,0 +1,10 @@
+package dms
+
+import (
+	"github.com/pivotal-cf/brokerapi"
+)
+
+// Provision implematation
+func (b *DMSBroker) Provision(instanceID string, details brokerapi.ProvisionDetails, asyncAllowed bool) (brokerapi.ProvisionedServiceSpec, error) {
+	return brokerapi.ProvisionedServiceSpec{}, nil
+}
