+++ pkg/services/dms/types.go
@@ -1,0 +1,13 @@
+package dms
+
+import (
+	"code.cloudfoundry.org/lager"
+	"github.com/huaweicloud/huaweicloud-service-broker/pkg/config"
+)
+
+// DMSBroker define
+type DMSBroker struct {
+	CloudCredentials config.CloudCredentials
+	Catalog          config.Catalog
+	Logger           lager.Logger
+}
