+++ pkg/services/obs/types.go
@@ -1,0 +1,13 @@
+package obs
+
+import (
+	"code.cloudfoundry.org/lager"
+	"github.com/huaweicloud/huaweicloud-service-broker/pkg/config"
+)
+
+// OBSBroker define
+type OBSBroker struct {
+	CloudCredentials config.CloudCredentials
+	Catalog          config.Catalog
+	Logger           lager.Logger
+}
