+++ pkg/services/obs/unbind.go
@@ -1,0 +1,10 @@
+package obs
+
+import (
+	"github.com/pivotal-cf/brokerapi"
+)
+
+// Unbind implematation
+func (b *OBSBroker) Unbind(instanceID, bindingID string, details brokerapi.UnbindDetails) error {
+	return nil
+}
