+++ pkg/services/rds/deprovision.go
@@ -1,0 +1,33 @@
+package rds
+
+import (
+	"fmt"
+
+	"github.com/huaweicloud/golangsdk/openstack/rds/v1/instances"
+	"github.com/pivotal-cf/brokerapi"
+)
+
+// Deprovision implematation
+func (b *RDSBroker) Deprovision(instanceID string, details brokerapi.DeprovisionDetails, asyncAllowed bool) (brokerapi.DeprovisionServiceSpec, error) {
+
+	// Init rds client
+	rdsClient, err := b.CloudCredentials.RDSV1Client()
+	if err != nil {
+		return brokerapi.DeprovisionServiceSpec{}, fmt.Errorf("create rds client failed. Error: %s", err)
+	}
+
+	// Log opts
+	b.Logger.Debug(fmt.Sprintf("deprovision rds instance opts: %s", instanceID))
+
+	// Invoke sdk
+	result := instances.Delete(rdsClient, instanceID)
+	if result.Err != nil {
+		return brokerapi.DeprovisionServiceSpec{}, fmt.Errorf("deprovision rds instance failed. Error: %s", err)
+	}
+
+	// Log result
+	b.Logger.Debug(fmt.Sprintf("deprovision rds instance success: %s", instanceID))
+
+	// Return result
+	return brokerapi.DeprovisionServiceSpec{IsAsync: false, OperationData: ""}, nil
+}
